+++ packages/reakit-test-utils/src/hover.ts
@@ -1,6 +1,4 @@
-import { fireEvent as domFireEvent } from "@testing-library/dom";
 import { fireEvent } from "./fireEvent";
-import { act } from "./act";
 
 type DocumentWithLastHovered = Document & {
   lastHovered?: Element | null;
@@ -30,10 +28,7 @@ export function hover(element: Element, options?: MouseEventInit) {
     fireEvent.mouseOut(lastHovered, options);
 
     if (!isElementWithinLastHovered) {
-      act(() => {
-        // fireEvent.mouseLeave would be the same as fireEvent.mouseOut
-        domFireEvent.mouseLeave(lastHovered, options);
-      });
+      fireEvent.mouseLeave(lastHovered, options);
     }
   }
 
@@ -43,10 +38,7 @@ export function hover(element: Element, options?: MouseEventInit) {
   fireEvent.pointerEnter(element, options);
   if (!disabled) {
     fireEvent.mouseOver(element, options);
-    act(() => {
-      // fireEvent.mouseEnter would be the same as fireEvent.mouseOver
-      domFireEvent.mouseEnter(element, options);
-    });
+    fireEvent.mouseEnter(element, options);
   }
   fireEvent.pointerMove(element, options);
   if (!disabled) {