+++ packages/reakit/src/Combobox/__examples__/ComboboxFetch/api.ts
@@ -48,7 +48,6 @@ const fruits = [
   "Kiwifruit",
   "Kumquat",
   "Lemon",
-  "lime",
   "Lime",
   "Longan",
   "Loquat",